Add files via upload
Decided to make the less complex by just making it a code where I just arrange strings using my own logic, and a little of help from the internet.
Currently, this file is containing an error, where, I am not able to input a string at the 0th index. So if I put the array length as 1, it will give an output immediately without even having the user put a string in.